Treat invalid port forwarding as administratively prohibited
A client may request a port outside of the range [0, 65535] as
the protocol uses a 32 bit integer to transfer the port number.
Any value outside of the standard range causes InetSocketAddress
to throw IllegalArgumentException, so trap the exception and fail
with the return code SSH_OPEN_ADMINISTRATIVELY_PROHIBITED.

We might be running with a SecurityManager enabled and be denied
access to the host's resolved IP address.  If this occurs we will
also now catch the SecurityException and convert it into the same
SSH_OPEN_ADMINISTRATIVELY_PROHIBITED return code.

git-svn-id: https://svn.apache.org/repos/asf/mina/sshd/trunk@881338 13f79535-47bb-0310-9956-ffa450edef68