Skip LoggingFilterOutputStream if logging is disabled
If the log channel isn't enabled at trace level when the streams
are being constructed, don't wrap them in LoggingFilterOuputStream
as it wastes CPU time to go through the unnecessary method dispatch
during write just to figure out we don't really need to be here.

Unfortunately this makes it impossible to dynamically enable the
trace logging, but given how verbose tracing would be on a running
server with multiple connections its likely not very valuable to
permit turning trace on dynamically without having a per-connection
filter to control which sessions are traced, and which are not.

git-svn-id: https://svn.apache.org/repos/asf/mina/sshd/trunk@890630 13f79535-47bb-0310-9956-ffa450edef68