Allocate the next ChannelOutputStream buffer sized to last buffer
If an output stream has just received a large burst of data from
the application, its buffer will have needed to grow beyond the
default buffer size of 512 bytes.  Its quite likely that there will
be even more data from the application, filling the next buffer
up to approximately the same size before we are forced to trigger
another flush.  So allocate the next buffer sized to hold the last
packet's complete length, or the data remaining if we had to split
the buffered packet to fit into the available window space.

Unfortunately most uses of createBuffer don't care about the size
of the buffer, as they are one-shot requests which are often within
the default size of 512 bytes.  Pass in 0 for these and allow them
to use the default 512 byte size.

git-svn-id: https://svn.apache.org/repos/asf/mina/sshd/trunk@890631 13f79535-47bb-0310-9956-ffa450edef68