Aggressively flush during write if the buffer is full
Instead of queuing up all data until the next flush() call made
by the application, queue data until we reach either the allowed
packet size or the current free window space, whichever limit is
encountered first.  As soon as buffer reaches either capacity,
send it immediately to the peer.

By sizing the buffer during write() to fit the available window
space we significantly reduce the chances that a buffer split and
copy is needed during flush().  However, a buffer split could still
occur if different threads are writing to two different streams on
the same channel (e.g. one thread is writing to the stdout stream
and a different thread is writing on the stderr stream).

During write() we give ourselves a credit equal to the size of
the last packet we successfully wrote out of this channel stream.
This allows the application producer to race ahead of the network
handshake by one packet and produce more data while waiting for
the handshake to free up window space.

git-svn-id: https://svn.apache.org/repos/asf/mina/sshd/trunk@890632 13f79535-47bb-0310-9956-ffa450edef68