SSHD-8: agent forwarding
The main change in this commit is the switch from mina as the transport mechanism to plain tomcat/apr sockets mostly because the mina apr transport does not support the unix domain sockets.  This allow to interact with the openssh agent on unix platforms, and this also secure the agent correctly.

git-svn-id: https://svn.apache.org/repos/asf/mina/sshd/trunk@891692 13f79535-47bb-0310-9956-ffa450edef68