SSHD-8: Secure the agent socket against race conditions
Binding a UNIX domain socket creates the socket file with the
current umask.  This may permit evil members of the caller's group
to connect to the socket before the listening process can mark it
as accessible only by the owner.

Instead, create a temporary directory that is accessible only to
the owner, and only after the location is locked down, create the
socket within it as a unique random name.

git-svn-id: https://svn.apache.org/repos/asf/mina/sshd/trunk@893394 13f79535-47bb-0310-9956-ffa450edef68