SSHD-73: Remove pointless newBuffer during flush
During flush we invoke newBuffer() to ensure that the Buffer we pass
down into the outgoing encode and transmit queue is independent of
the Buffer we continue to use for subsequently written data.

However, there is no need to invoke newBuffer() again at the end
of the flush method.

If the method completes normally, the loop broke out because the
bufferLength member is equal to 0, which means the current Buffer
instance is in use only by ourselves, and is empty.  We don't need
to allocate yet another empty Buffer to replace it.

If the method crashes due to an exception, we have only two cases
to worry about:

  case 1:  before the newBuffer call

  In this case, the only state changes we have done is to write
  the current buffer length into the existing Buffer instance.
  Nobody else has our Buffer.  If we crash with a RuntimeException
  during this phase and come back here again during a future write
  we'll just overwrite the same section with a new length.

  case 2:  after the newBuffer call

  After the newBuffer invocation we could crash while trying to
  send the old Buffer out.  We've already replaced the Buffer with
  a fresh one, so all that happens is the old Buffer gets garbage
  collected, possibly causing a gap in the application data stream.

So this use of newBuffer during the finally is unnecessary and
is not needed to ensure our future instance state is sane if an
exception occurs.

git-svn-id: https://svn.apache.org/repos/asf/mina/sshd/trunk@897372 13f79535-47bb-0310-9956-ffa450edef68