Correctly compute the lastSize member in ChannelOutputStream
If we have to split the current Buffer, because its too large for
the session's window or packet size, don't give ourselves a credit
equal to the left over portion.  Instead use the actual number of
bytes we transmit inside of flush.

However, because we know the new Buffer needs to hold at least
a certain size, we can allocate it in advance to that target,
to avoid needing to grow it while doing the split.

git-svn-id: https://svn.apache.org/repos/asf/mina/sshd/trunk@897373 13f79535-47bb-0310-9956-ffa450edef68