SSHD-73: Avoid infinite loop when the window is full
If the window has no more free space, we can't flush any more data
until we receive a window adjust from the peer and get more space.

If we still have data to write, but there is no data in our local
buffer, we can't rely on flush() to wait for the window to get more
free space.  Calling flush would do nothing, and we'd spin in a hard
loop without releasing the processor until the window adjust came in.

Instead of spinning, fall sleep on the Window until there is at
least one byte free for us to write into.

The only way for this case to arise is for both lastSize and
bufferLength to be equal to 0, and for the Window to have no space.
This can only happen if the application has written sufficient data
to one OutputStream (stdout or SSH_MSG_CHANNEL_DATA) to cause the
Window to have no space, and then it performs the first write on
the other OutputStream (stderr or SSH_MSG_CHANNEL_EXTENDED_DATA).

This can happen if the first stream throws an exception, and the
application catches it and tries to print the stack trace to the
second stream.

When the 2nd output stream starts its first write, both lastSize and
bufferLength are 0 as we have never sent a packet from this stream.
We spin hard in the write-flush loop, and never actually fall asleep
to wait for the window adjust.

git-svn-id: https://svn.apache.org/repos/asf/mina/sshd/trunk@897374 13f79535-47bb-0310-9956-ffa450edef68