SSHD-19: Avoid NullPointerException when sftp is disabled
If a subsystem is disabled the server won't be able to create
the subsystem command object and initialize it into the channel.
A SSH_MSG_CHANNEL_FAILURE is returned to the client, but the
client may have already started to send data for this command.

In the case of the sftp there is an initial data packet
sent by the client immediately behind the channel request,
so the ChannelSession needs to consume this requset even if it
no subsystem command was created.  Eventually the client will
receive the SSH_MSG_CHANNEL_FAILURE that was sent, and tear down
the connection with a failure message.

git-svn-id: https://svn.apache.org/repos/asf/mina/sshd/trunk@1031886 13f79535-47bb-0310-9956-ffa450edef68