clear response buffer with unsupported extension
As it is, if a client requests to use an extension which is not
supported by the server, the server is supposed to send
`SSH_FX_OP_UNSUPPORTED`. It does send that, but it doesn't clear the
`Buffer` object before sending the response. So, all the clients who
are requesting an unsupported extension are impacted with invalid
bytestream of response data and the channel becomes unusable.

This patch simply clears the buffer with `prepareReply` function before
sending the response.