refactor sendHandle to call prepareReply by default
All the callsites of `sendHandle` overloaded functions invoke them by
clearing the buffer explicitly, with `prepareReply`. This establishes an
implicit contract which is not enforced by any tools. So, its quite easy
to miss that and end up introducing bugs.

This patch changes all the callsites of `sendHandle` functions to not to
explicitly invoke `prepareReply` and changes `sendHandle` function to
call that function on the passed buffer object.