[SSHD-929] Create separate listener for remove directory.
Right now listener for deletion of file and directory is common. This
makes it difficult to find whether directory / file has been removed.

The path passed to `removed` function in SftpEventListener will not tell
if the deleted path was a directory or file. `Files.isDirectory(path)`
will return false, as the directory doesn't exist.

This patch will have a fix to create one more listener entry for directory removal