[SSHD-1050] Fix race conditions with early access to AuthFuture
Make ClientSessionImpl.authFuture volatile since it may be accessed by
different threads without synchronization.

Remove the initial dummy AuthFuture that served only to record an early
exception, and to provide a correct session state in waitFor() when KEX
was done but auth() not called yet.

Record exceptions from signalAuthFailure() in a list if there's no
AuthFuture yet. Introduce a lock for guarding the authFuture field and
this new list of exceptions together.

The access in updateCurrentSessionState() must remain unsynchronized
because it occurs while the futureLock is held. Synchronizing there
on errorLock might introduce a lock inversion between that access in
updateCurrentSessionState() (lock order futureLock-errorLock), but
the call to authService.auth() inside ClientSessionImpl.auth() has the
potential to also access the futureLock is there was a previous future,
which might give the lock order errorLock-futureLock.

Listeners futures are always called outside the future's lock, so even
if a listener calls auth() or signalAuthFailure(), this will
not cause a lock inversion.