SSHD-1166 - Support creating signed OpenSSH Certificates
Add a new OpenSshCertificateBuilder for creating new certificates.

Add an enum for the certificate type (user/host), and a type
CertificateOption for the "critical options" and "extensions" that
can be stored in a certificate. Add new put/get methods to Buffer for
this new CertificateOptions type.

Fix a subtle bug observed in KeyUtils mapping data, where order
matters for discovery in a few places.

Add javadoc to the OpenSshCertificate methods.

Make tests also verify that signatures on generated certificates can
be verified, and add a new test that compares generated certificates
against ones that were created by OpenSSH.