[SSHD-1197] Fix race condition in KEX
There was a window in AbstractSession.requestNewKeyExchange()
during which the KEX state was set already to INIT, but the
caller's proposal not set yet. If the peer had also decided to
start a new key exchange, it was possible that the peer's KEX_INIT
message arrived and was handled in that window and then proceeded
with wrong or uninitialized data.

KEX negotiation must start only when both proposals are indeed
available. Thus wait when having received a KEX_INIT from the peer
until our own proposal has been prepared before continuing with the
negotiation.

There's a new CoreModuleProperties.KEX_PROPOSAL_SETUP_TIMEOUT property
for controlling the timeout for this waiting. Since it's used at a
moment when there is already a thread running to prepare our proposal,
we should never have to wait for long; the property should be set to a
reasonably high value such that the timeout is never hit. A minimum
value for this timeout of 5 seconds is enforced.