[SSHD-1188] SFTP: add a test for renaming a file
If the target already exists, a plain rename must fail. For SFTP >= V5,
a rename with CopyMode.Overwrite shall succeed. Note that the behavior
with CopyMode.Atomic is platform dependent; it's mapped internally to
java.nio.file.Files.move() with StandardCopyOption.ATOMIC_MOVE, for
which the javadoc says

"If the target file exists then it is implementation specific if
the existing file is replaced or this method fails by throwing an
IOException."

This is fine; the SFTP RFC draft, version 13, allows this behavior for
atomic renames.[1]

[1] https://datatracker.ietf.org/doc/html/draft-ietf-secsh-filexfer-13#section-8.3