Allow using an SSH agent without agent forwarding
It's quite possible to use an SSH agent without agent forwarding.
Apache MINA sshd should not require ChannelForwardingFactories being
implemented by the SshAgentFactory. If it does, fine; if it does not,
it's also fine and agent forwarding will simply not be available.

Agent forwarding comes with a well-known security risk: root on the
server could abuse the forwarded agent. Using the more modern
ProxyJump feature it's possible to connect through jump hosts without
agent forwarding (and without needing to put keys onto the jump host).

So give users the option of providing an SshAgentFactory that only
implements createClient(), but that returns null or an empty list from
getChannelForwardingFactories().