[SSHD-1218] Provide the IdentityAgent setting to the SshAgentFactory
Set an attribute on the ClientSession. Since commit 14b4ec07, the
factory gets passed the session when creating an SshAgent and can thus
obtain this value from the session. It needs to know this value to be
able to connect to the correct SSH agent.

Also change the handling of SSH agents when IdentitiesOnly is set. Do
not hard-code skipping an agent completely if the value is true. The
value is also available via the ClientSession; hence just let the
factory decide what it should do in that case. This is more flexible
for users of the library.