[SSHD-1220] Cache SFTP attributes on SftpPaths
This greatly reduces the number of (L)STAT calls made to get SFTP
attributes. Also, listing directories in reply to SSH_FXP_OPENDIR
and SSH_FXP_READDIR can in this way still go through the normal
`java.nio.file.FileSystem` abstraction. There is no need anymore to
bypass the file system to resolve SSHD-1217.

The attributes cached on the `Path` objects returned from the
`DirectoryStream` are used only for listing the directory. This
is a similar optimization as was done in the Java libraries for
a `FileTreeWalker` operating on an Unix file system, which may use
cached attributes via the internal interface
`sun.nio.fs.BasicFileAttributesHolder`.

Caching is also enabled selectively in the up-front checks for file
accessibility, which also caused repeated (L)STAT calls. Rewrite
`IoUtils.checkFileExists()` to load attributes only once.

The new operations for caching attributes on `SftpPath` instances are
not intended to be used by client code. For access in the SFTP
implementation itself, paths are now always created as `SftpPathImpl`
instances, which inherit from `SftpPath` and publish a safe
`withAttributeCache()` method. Theoretically client code could
still access these methods, but then it's clear for everyone that these
are internal operations.