[SSHD-1221] Support key constraints when adding a key to the SSH agent
Change the API of SshAgent.addKeyToAgent() to be able to pass key
constraints. Provide default implementations for the documented
OpenSSH constraints: confirm, lifetime, and the sk-provider extension.

Handle them in the AbstractAgentProxy implementation. Note that
AbstractAgentProxy currently implements addKeyToAgent() only for
OpenSSH. Users wanting to support a different agent may have to
implement different constraints by extending the provided base classes,
and may have to provide their own SshAgent (and a factory for it)
implementing a different protocol for addKeyToAgent().