SFTP: support file timestamps with sub-millisecond resolution
Refactor the reading and writing of FileTimes to avoid manual nanosecond
calculations. Go through Instant instead. On writing, this may saturate,
but for times that far in the past or future nanos are irrelevant. On
reading, drop nanos if the value would be outside the range of Instant.

The advantage of doing it this way is that we can actually report and
recover sub-millisecond timestamps (for SFTP version >= 4). Newer Java
does report file timestamps with the OS resolution (100ns on Windows/
NTFS, may be even lower on other OSes depending on the file system).