[SSHD-1254] Client side implementation of host-bound pubkey auth
Provide a KexExtensionParser for it; make the default client KEX
extension handler understand the "publickey-hostbound@openssh.com"
extension, and use it in UserAuthPublicKey.

The extension changes the auth name in the SSH_MSG_USERAUTH_REQUEST
from "pubkey" to "publickey-hostbound-v00@openssh.com" and includes
the server's public key in the message and its signature.

Includes a container test using publickey authentication to log in to
an OpenSSH 9.0 server, verifying that the Apache MINA sshd client got
the KEX extension and checking the server log from the container to
verify that the server got a host-based signature.

See https://github.com/openssh/openssh-portable/blob/807be686/PROTOCOL#L347