[SSHD-1257] ChannelSession: don't flush out stream if already closed
Since OutputStream.close() is a no-op when the stream is already closed
and does flush if not, use close() instead of flush().

This use in ChannelSession is the last use of this OutputStream. It
may get closed again when the ChannelSession itself closes, but that
is then just a no-op.