KEX: change output encoding atomically
Ensure the `encodeLock` is held for sending the `SSH_MSG_NEWKEYS`
message and changing the output encoding.

For the input encoding a similar change is not needed since the
`decodeLock` is held anyway.

Strictly speaking acquiring the `encodeLock` should not be necessary
since any other potentially concurrent writes should be queued because
the KEX state is not yet `DONE`.