Fine-grained synchronization in ChannelAsyncOutputStream
Remove the broad locking on the stream instance's monitor. Holding
such broad locks during write operations may easily deadlock. Instead
guard only some state fields, and make sure these are never locked
during write operations, or when futures are fulfilled, which may
notify arbitrary listeners, which might trigger writes.

The idea of this ChannelAsyncOutputStream is to write as much of a
buffer as possible as individual packets. When there is not enough
window space, stop writing and let the write be resumed asynchronously
once the window has been enlarged; then write the rest of the buffer.

The implementation treats the buffer to be written as a token: while
it is being processed, the pendingWrite field is null. This ensures
that always only one thread is processing the buffer to write the next
chunk. Otherwise it might be possible that an asynchronous window
adjustment tried to process the buffer while it already was being
written.

Closing a ChannelAsyncOutputStream gracefully is tricky and may not
always be possible, depending on the reason for closing it. If channel
window adjustments are not handled anymore, writing the last pending
buffer in full may not be possible if the stream is waiting for more
window space. In that case, the stream may close with some unwritten
data, but that is unavoidable. (The previous code had a similar problem;
see SSHD-1261; this implementation should also solve that issue.)