[SSHD-1264] Add a test case
Use a CentOS container running OpenSSH 7.4 configured to re-key every
second. Connect with an Apache MINA sshd client and write some data
every second for 20 seconds. Check that we got multiple key exchanges.

Install a ServerKeyVerifier that fails if the host key on re-keying is
different from the key presented in the initial KEX.

Note that the problem is not reproducible with newer OpenSSH (as
available for instance for Alpine containers).