Fix the Netty connection back-end
Running the tests with Netty frequently resulted in failures due to
connection time-outs. The root cause of this was a race condition
in NettyIoConnector: if the connection succeeded and the NettyIoSession
was initialized via channelActive() before the connector had set
the DefaultIoConnectFuture attribute on the channel, then that future
would never be signalled.

Rewrite the NettyIoConnector to not use a global Bootstrap for all
connections. Use a new and dedicated Bootstrap per connection. This
makes is possible to use Bootstrap.attr() to have the channel attribute
be set early on, before even the connection is attempted. For the
context AttributeRepository, one then doesn't even need an attribute;
it's available directly.

Also don't set SO_BACKLOG in the connector; this socket option only
makes sense for server sockets, i.e., in the acceptor.

Enable running the sshd-core tests also with Netty.